Moved tool crash tests to the `tools` module
