+++ frontend/wasm/src/editor_api.rs
@@ -435,23 +435,3 @@ impl Drop for JsEditorHandle {
 		EDITOR_INSTANCES.with(|instances| instances.borrow_mut().remove(&self.editor_id));
 	}
 }
-
-#[cfg(test)]
-mod layer_delete_bug_tests {
-	use super::*;
-
-	use editor::misc::test_utils::EditorTestUtils;
-	use editor::viewport_tools::tool::ToolType;
-
-	#[test]
-	fn should_not_crash_when_layer_is_deleted_while_using_a_tool() {
-		editor::communication::set_uuid_seed(0);
-		let mut test_editor = Editor::new();
-
-		test_editor.select_tool(ToolType::Pen);
-		test_editor.lmb_mousedown(0.0, 0.0);
-		test_editor.move_mouse(100.0, 100.0);
-
-		test_editor.handle_message(DocumentMessage::DeleteSelectedLayers);
-	}
-}
